Extract handlers.
