+++ src/dispatcher.rs
@@ -2,7 +2,7 @@
 //!
 
 use crate::{
-    cpu::{Rv32c, Rv32f, Rv32i, Rv32m},
+    handlers::{Rv32cHandler, Rv32fHandler, Rv32iHandler, Rv32mHandler},
     tobits::ToBits,
 };
 
@@ -13,14 +13,14 @@ pub trait Rv32iDispatcher {
     type Item;
 
     /// Decodes the input word to an RV32I instruction and dispatches it to a handler.
-    fn dispatch(&mut self, code: u32) -> <Self as Rv32i>::Item
+    fn dispatch(&mut self, code: u32) -> <Self as Rv32iHandler>::Item
     where
-        Self: Rv32i;
+        Self: Rv32iHandler;
 }
 
 impl<T, U> Rv32iDispatcher for T
 where
-    T: Rv32i + Rv32i<Item = U>,
+    T: Rv32iHandler + Rv32iHandler<Item = U>,
 {
     type Item = U;
 
@@ -94,14 +94,14 @@ pub trait Rv32imDispatcher {
     type Item;
 
     /// Decodes the input word to an RV32IM instruction and dispatches it to a handler.
-    fn dispatch(&mut self, code: u32) -> <Self as Rv32i>::Item
+    fn dispatch(&mut self, code: u32) -> <Self as Rv32iHandler>::Item
     where
-        Self: Rv32i + Rv32m;
+        Self: Rv32iHandler + Rv32mHandler;
 }
 
 impl<T, U> Rv32imDispatcher for T
 where
-    T: Rv32i + Rv32i<Item = U> + Rv32m + Rv32m<Item = U>,
+    T: Rv32iHandler + Rv32iHandler<Item = U> + Rv32mHandler + Rv32mHandler<Item = U>,
 {
     type Item = U;
 
@@ -183,14 +183,14 @@ pub trait Rv32icDispatcher {
     type Item;
 
     /// Decodes the input word to an RV32IC instruction and dispatches it to a handler.
-    fn dispatch(&mut self, code: u32) -> <Self as Rv32i>::Item
+    fn dispatch(&mut self, code: u32) -> <Self as Rv32iHandler>::Item
     where
-        Self: Rv32i + Rv32c;
+        Self: Rv32iHandler + Rv32cHandler;
 }
 
 impl<T, U> Rv32icDispatcher for T
 where
-    T: Rv32i + Rv32i<Item = U> + Rv32c + Rv32c<Item = U>,
+    T: Rv32iHandler + Rv32iHandler<Item = U> + Rv32cHandler + Rv32cHandler<Item = U>,
 {
     type Item = U;
 
@@ -312,14 +312,19 @@ pub trait Rv32imcDispatcher {
     type Item;
 
     /// Decodes the input word to an RV32IMC instruction and dispatches it to a handler.
-    fn dispatch(&mut self, code: u32) -> <Self as Rv32i>::Item
+    fn dispatch(&mut self, code: u32) -> <Self as Rv32iHandler>::Item
     where
-        Self: Rv32i + Rv32m + Rv32c;
+        Self: Rv32iHandler + Rv32mHandler + Rv32cHandler;
 }
 
 impl<T, U> Rv32imcDispatcher for T
 where
-    T: Rv32i + Rv32i<Item = U> + Rv32m + Rv32m<Item = U> + Rv32c + Rv32c<Item = U>,
+    T: Rv32iHandler
+        + Rv32iHandler<Item = U>
+        + Rv32mHandler
+        + Rv32mHandler<Item = U>
+        + Rv32cHandler
+        + Rv32cHandler<Item = U>,
 {
     type Item = U;
 
@@ -449,14 +454,19 @@ pub trait Rv32imfDispatcher {
     type Item;
 
     /// Decodes the input word to an RV32IMF instruction and dispatches it to a handler.
-    fn dispatch(&mut self, code: u32) -> <Self as Rv32i>::Item
+    fn dispatch(&mut self, code: u32) -> <Self as Rv32iHandler>::Item
     where
-        Self: Rv32i + Rv32m + Rv32f;
+        Self: Rv32iHandler + Rv32mHandler + Rv32fHandler;
 }
 
 impl<T, U> Rv32imfDispatcher for T
 where
-    T: Rv32i + Rv32i<Item = U> + Rv32m + Rv32m<Item = U> + Rv32f + Rv32f<Item = U>,
+    T: Rv32iHandler
+        + Rv32iHandler<Item = U>
+        + Rv32mHandler
+        + Rv32mHandler<Item = U>
+        + Rv32fHandler
+        + Rv32fHandler<Item = U>,
 {
     type Item = U;
 
@@ -579,21 +589,21 @@ pub trait Rv32imfcDispatcher {
     type Item;
 
     /// Decodes the input word to an RV32IMFC instruction and dispatches it to a handler.
-    fn dispatch(&mut self, code: u32) -> <Self as Rv32i>::Item
+    fn dispatch(&mut self, code: u32) -> <Self as Rv32iHandler>::Item
     where
-        Self: Rv32i + Rv32m + Rv32f + Rv32c;
+        Self: Rv32iHandler + Rv32mHandler + Rv32fHandler + Rv32cHandler;
 }
 
 impl<T, U> Rv32imfcDispatcher for T
 where
-    T: Rv32i
-        + Rv32i<Item = U>
-        + Rv32m
-        + Rv32m<Item = U>
-        + Rv32f
-        + Rv32f<Item = U>
-        + Rv32c
-        + Rv32c<Item = U>,
+    T: Rv32iHandler
+        + Rv32iHandler<Item = U>
+        + Rv32mHandler
+        + Rv32mHandler<Item = U>
+        + Rv32fHandler
+        + Rv32fHandler<Item = U>
+        + Rv32cHandler
+        + Rv32cHandler<Item = U>,
 {
     type Item = U;
 
@@ -759,7 +769,7 @@ where
 
 #[cfg(test)]
 mod test {
-    use crate::cpu::{Rv32c, Rv32f, Rv32i, Rv32m};
+    use super::{Rv32cHandler, Rv32fHandler, Rv32iHandler, Rv32mHandler};
     use crate::reg::Reg;
 
     // Base RV32I - generated by parse.py in riscv-opcodes.
@@ -979,7 +989,7 @@ mod test {
 
     struct Checker;
 
-    impl Rv32i for Checker {
+    impl Rv32iHandler for Checker {
         type Item = &'static str;
 
         fn illegal(&mut self, _ins: u32) -> Self::Item {
@@ -1194,7 +1204,7 @@ mod test {
         assert_eq!("xori", c.dispatch(enc(MATCH_XORI, MASK_XORI)));
     }
 
-    impl Rv32m for Checker {
+    impl Rv32mHandler for Checker {
         type Item = &'static str;
 
         fn mul(&mut self, _rd: Reg, _rs1: Reg, _rs2: Reg) -> Self::Item {
@@ -1245,7 +1255,7 @@ mod test {
         assert_eq!("remu", c.dispatch(enc(MATCH_REMU, MASK_REMU)));
     }
 
-    impl Rv32c for Checker {
+    impl Rv32cHandler for Checker {
         type Item = &'static str;
 
         fn c_addi4spn(&mut self, _rdp: Reg, _imm: u32) -> Self::Item {
@@ -1397,7 +1407,7 @@ mod test {
         assert_eq!("c.xor", c.dispatch(enc(MATCH_C_XOR, MASK_C_XOR)));
     }
 
-    impl Rv32f for Checker {
+    impl Rv32fHandler for Checker {
         type Item = &'static str;
 
         fn flw(&mut self, _rd: Reg, _rs1: Reg, _iimm: u32) -> Self::Item {
